Move sync code into util class
